requirements.txt
Flask
requests
psycopg2-binary
sqlalchemy
gunicorn
azure-identity
azure-keyvault-secrets
python-dotenv